+++ src/lib/types/cache/cache-types.ts
@@ -1,3 +1,5 @@
+import { Intention } from "../ai-response";
+
 export enum RoleEnum {
     USER = 'user',
     ASSISTANT = 'assistant'
@@ -6,4 +8,5 @@ export interface ChatMessage {
   // waId: string;
   role: RoleEnum;
   content: string;
+  intention?: Intention;
 }