+++ src/modules/reservations/service/helpers/get-missing-fields.helper.ts
@@ -1,0 +1,7 @@
+import { DeleteReservation } from "src/lib";
+
+export function getMissingFields(obj: DeleteReservation): string[] {
+    return Object.entries(obj)
+        .filter(([_, value]) => value === null || value === undefined || value === '')
+        .map(([key]) => key);
+}