+++ src/modules/reservations/service/intention/create-reservation.strategy.ts
@@ -36,6 +36,7 @@ export class CreateReservationStrategy implements IntentionStrategyInterface {
             case TemporalStatusEnum.IN_PROGRESS:
                 return {reply: await this.aiService.getMissingData(response.missingFields, history)};
             case TemporalStatusEnum.COMPLETED:
+                await this.cacheService.clearHistory(mockedData.waId);
                 return {reply: await this.aiService.reservationCompleted(response.reservationData, history)};
             default:
                 this.logger.warn(`Estado de reserva inesperado: ${response.status}`);